+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.3.0 (Sep 18, 2017)
+
+- Added 4 new scope properties: `totalDays`, `totalHours`, `totalMinutes` and `totalSeconds` (#1).
+- Added a watch to activate timer after creation (#2).
 
 ## 0.2.0 (Apr 28, 2017)
 
@@ -13,14 +17,13 @@ Vue.use(VueCountdown);
 Vue.component('countdown', VueCountdown);
 ```
 
-
 ## 0.1.1 (Apr 27, 2017)
 
 - Improved countdown progress.
-
 
 ## 0.1.0 (Apr 10, 2017)
 
 - Supports 4 props (options): `autoStart`, `interval`, `time` and `tag`.
 - Supports 2 methods: `start`, `stop`.
 - Supports 3 events: `countdownstart`, `countdownprogress` and `countdownend`.
+- Supports 4 scope properties: `days`, `hours`, `minutes` and `seconds`.